Backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.10.5
pydantic-settings==2.7.1

# Redis & Caching
redis==5.2.1
hiredis==3.0.0

# Background Tasks
celery==5.4.0
flower==2.0.1

# Validation & Serialization
pydantic==2.10.5

# File Handling
openpyxl==3.1.5
pandas==2.2.3

# Date/Time
python-dateutil==2.9.0
